add test for partial update
